Refactor actions and move model
