+++ .gitignore
@@ -22,3 +22,4 @@ node_modules/
 /actions/types/
 /cli/artifacts
 /cli/types
+yarn.lock
